Add alternative implementation based on OpenFCI and C++ interface
A faster and more accurate implementation based on OpenFCI has been
added.  The makefile can automatically download and build the
dependencies needed by OpenFCI.  The previous implementation (`am`)
remains the default in the makefile.

A C++ interface has been added for convenience.  This interface is a
header-only wrapper around the existing C interface.