Major revamp of the library
The structure and interface of the library has been changed.

A new, backward-incompatible "v2" interface has been introduced, which
groups several matrix element calculations into a single batch (called
"tabulation request").

When the user requests a set of matrix elements, the library writes a
temporary file containing the indices of the requested matrix elements.
It then spawns a child process that handles the actual calculations.
The child process ("tabulation provider") reads the file, writes the
values of the matrix elements into the same file, and exits once it
completes.  The library then reads in the results and returns them to
the user.

Additionally, the OpenFCI interface has been moved into a separate
project of its own (clh2-openfci).